refactor : pages 디렉토리 layout 위치 수정
디렉토리 일관성 확보
