+++ src/global.css
@@ -17,7 +17,7 @@
 
 @import './pages/main-page/main-page.css';
 @import './pages/search-page/search-page.css';
-@import './pages/layout/layout.css';
+@import './pages/common/layout/layout.css';
 
 body {
   background-color: var(--color-bluegray-100);
