+++ src/main.ts
@@ -1,4 +1,4 @@
-import Layout from './pages/layout/Layout';
+import Layout from './pages/common/layout/Layout';
 import { initRouter } from './route/router';
 
 addEventListener('load', () => {
