+++ src/pages/common/layout/Layout.ts
@@ -1,6 +1,6 @@
-import Footer from '../../component/layout/footer/Footer';
-import Header from '../../component/layout/header/Header';
-import { renderContent } from '../../route/router';
+import Footer from '../../../component/layout/footer/Footer';
+import Header from '../../../component/layout/header/Header';
+import { renderContent } from '../../../route/router';
 
 class Layout {
   #container;
